A PR that's an issue
Opening this as a pull request because you don't have issues enabled; one of this package's unit tests breaks with the API changes in the new urltools release (I have...literally no idea why). But if you install urltools from github and run the tests, you'll see what I mean. I can't release until this is fixed, so it'd be great if y'all could work out what's going on (and let me know if it's an issue at my end).